src/optimization/GOTAcc/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gotacc"
dynamic = ["version"]
description = "General Optimization Toolkit for Accelerator Applications"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.10"
license = { text = "MIT" }

authors = [
  { name = "Haoran Zhang" }
]

keywords = [
  "accelerator",
  "optimization",
  "bayesian-optimization",
  "multi-objective-optimization",
  "beam-dynamics",
  "epics"
]

classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Scientific/Engineering :: Physics",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Operating System :: OS Independent"
]

dependencies = [
  "numpy>=1.24",
  "scipy>=1.10",
  "matplotlib>=3.7",
  "torch>=2.2",
  "gpytorch>=1.11",
  "botorch>=0.10"
]

[project.optional-dependencies]
yaml = [
  "PyYAML>=6.0"
]
gui = [
  "PyQt5>=5.15"
]
epics = [
  "pyepics>=3.5"
]
full = [
  "PyYAML>=6.0",
  "PyQt5>=5.15",
  "pyepics>=3.5"
]
dev = [
  "pytest>=8.0",
  "black>=24.0",
  "isort>=5.13"
]

[project.urls]
Homepage = "https://github.com/ustc-zhr/GOTAcc"
Repository = "https://github.com/ustc-zhr/GOTAcc"
Issues = "https://github.com/ustc-zhr/GOTAcc/issues"

[project.scripts]
gotacc-run = "gotacc.runners.run_cli:main"
gotacc-gui = "gotacc.gui.main:main"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["gotacc*"]

[tool.setuptools.package-data]
gotacc = [
  "configs/pv_lists/*.json",
  "configs/yaml_cfg/*.yaml",
  "configs/yaml_cfg/*.yml",
  "gui/*.json",
  "gui/ui/*.ui"
]

[tool.setuptools.dynamic]
version = { attr = "gotacc.version.__version__" }

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
